Add files via upload
I used a virtual environment as a part of all the other work I do and to be able to debug in some directory that is not clustered, but it should not be mandatory to use. I tried the website on my local host, and it compiles well as I intended. It just needs all the imported modules with specific versions of Python and CSS colour code and a bin that runs all the activation scripts.